split Objectpage into planetpage, moonpage, and bodypage so that I can use the router from Main.jsx to specify which one to display
