+++ src/components/Main.jsx
@@ -2,7 +2,9 @@ import { Route, Routes } from "react-router-dom";
 
 import Home from './Home'
 import CollectionPage from './CollectionPage'
-import ObjectPage from "./ObjectPage";
+import PlanetPage from "./PlanetPage";
+import MoonPage from "./MoonPage"
+import BodyPage from "./BodyPage"
 
 import NotFound from "./NotFound";
 
@@ -12,11 +14,11 @@ const Main = () => {
         <Routes>
             <Route exact path="/" element={<Home />} />
             <Route exact path={`/planets`} element={<CollectionPage />} />
-            <Route exact path={`/planets/:id`} element={<ObjectPage />} />
+            <Route exact path={`/planets/:id`} element={<PlanetPage />} />
             <Route exact path={`/moons`} element={<CollectionPage />} />
-            <Route exact path={`/moons/:id`} element={<ObjectPage />} />
+            <Route exact path={`/moons/:id`} element={<MoonPage />} />
             <Route exact path={`/bodies`} element={<CollectionPage />} />
-            <Route exact path={`/bodies/:id`} element={<ObjectPage />} />
+            <Route exact path={`/bodies/:id`} element={<BodyPage />} />
             <Route exact path="/:id" element={<NotFound />} /> 
         </Routes>
     )
