--- src/components/ObjectPage.jsx
@@ -1,71 +1,0 @@
-import { Link, useLocation} from "react-router-dom";
-import useApi from "../useApi"
-
-const ObjectPage = () => {
-
-    let location = useLocation()
-    let path = location.pathname
-
-    const { data:object } = useApi(path)
-
-    if ( path.includes("planets") && !object) {return <h3>Loading Planet</h3>}
-    if ( path.includes("moons") && !object) {return <h3>Loading Moon</h3>}
-    if ( path.includes("bodies") && !object) {return <h3>Loading Body</h3>}
-
-    if ((typeof object) !== "object") {return <h3>{object}</h3>}
-
-
-    if (path.includes("planets")) {
-        return (
-             <div className="detail-div">
-            <p className="detail-div-title-text">{object.name}</p>
-            <p className="mass-text">Mass: {object.mass}</p>
-            <p className="diameter-text">Diameter: {object.diameter}</p>
-            <p className="dist_from_sun-text">Distance from the Sun:{object.dist_from_sun}</p>
-            <p className="terrain-text">Terrain: {object.terrain}</p>
-            <p className="water-text">Water Present: {object.has_water}</p>
-            <p className="alien-text">Aliens: {object.has_aliens}</p>
-            <p className="surf-temp-text">Average Surface Temperature: {object.surface_temperature}</p>
-            <p className="gravity-text">Gravity: {object.gravity}</p>
-            <Link to="/planets" className="back-btn">Back</Link>
-        </div>
-        )
-    }
-
-    if (path.includes("moons")) {
-        return (
-            <div className="detail-div">
-                <p className="detail-div-title-text">{object.name}</p>
-                <p className="home-planet-text">Orbiting around {object.home_planet}</p>
-                <p className="mass-text">Mass: {object.mass}</p>
-                <p className="diameter-text">Diameter: {object.diameter}</p>
-                <p className="dist_from_planet-text">Distance from {object.home_planet}:{object.dist_from_planet}</p>
-                <p className="water-text">Water Present: {object.has_water}</p>
-                <p className="surf-temp-text">Surface Temperature: {object.surface_temperature}</p>
-                <p className="gravity-text">Gravity: {object.gravity}</p>
-                <Link to="/moons" className="back-btn">Back</Link>
-            </div>
-        )
-    }
-
-    if ( path.includes("bodies")) {
-        return (
-            <div className="detail-div">
-                <p className="detail-div-title-text">{object.name}</p>
-                <p className="mass-text">Mass: {object.mass}</p>
-                <p className="diameter-text">Diameter: {object.diameter}</p>
-                <p className="terrain-text">Terrain: {object.terrain}</p>
-                <p className="water-text">Water Present: {object.has_water}</p>
-                <p className="alien-text">Aliens: {object.has_aliens}</p>
-                <p className="surf-temp-text">Average Surface Temperature: {object.surface_temperature}</p>
-                <p className="gravity-text">Gravity: {object.gravity}</p>
-                <Link to="/bodies" className="back-btn">Back</Link>
-            </div>
-        )
-    }
-    return (
-        <div>invalid url</div>
-    ) 
-}
-
-export default ObjectPage
